+++ src/h3lib/include/coordijk.h
@@ -31,8 +31,16 @@
 #define COORDIJK_H
 
 #include "geoCoord.h"
-#include "h3api.h"
 #include "vec2d.h"
+
+/** @struct CoordIJK
+ * @brief IJK hexagon coordinates
+ */
+typedef struct {
+    int i;  ///< i component
+    int j;  ///< j component
+    int k;  ///< k component
+} CoordIJK;
 
 /** @brief CoordIJK unit vectors corresponding to the 7 H3 digits.
  */
@@ -93,5 +101,6 @@ void _ijkRotate60ccw(CoordIJK* ijk);
 void _ijkRotate60cw(CoordIJK* ijk);
 Direction _rotate60ccw(Direction digit);
 Direction _rotate60cw(Direction digit);
+int ijkDistance(const CoordIJK* a, const CoordIJK* b);
 
 #endif